chore: Delete unused boiler-plate code/styles so that I can start fresh with styling
`MainLayout.razor.css` is left empty so that `DotNetCommunity.DotNetWatchApp.styles.css` is still generated and we don't get a 404.

`app.css` keeps validation and blazor error rules., and one peculiar rule to remove outline from focused h1. Blazor adds `tabindex=-1` to the `<h1>` which can cause it to be focused and have a border with the browser styling. I do not understand why Blazor adds this, so I intend to circle back and document this rule when I do.